# api/openapi.yaml
openapi: 3.1.0
info:
  title: HeatCheck Public API
  version: 1.0.0
  description: |
    Production API for HeatCheck challenges, proof clips, OpenAI-assisted
    verification, paid participation, voting, moderation, account lifecycle,
    and copyright workflows. All timestamps use RFC 3339 UTC values.
servers:
  - url: https://heatcheck.dogi.watch
    description: Production
  - url: http://localhost:8080
    description: Local development
...
security: []
paths:
...
  /v1/policies/{kind}:
    parameters:
      - $ref: "#/components/parameters/PolicyKind"
    get:
      tags: [Policies]
      operationId: getCurrentPolicy
      summary: Get a current policy document
      responses:
        "200":
          description: Current policy document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyEnvelope"
        "404":
          $ref: "#/components/responses/Error"
...
  /v1/users/{userID}/block:
    parameters:
      - $ref: "#/components/parameters/UserID"
    put:
      tags: [Safety]
      operationId: blockUser
      summary: Block a user
      security:
        - bearerAuth: []
      responses:
        "204":
          description: User blocked.
        "404":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Safety]
      operationId: unblockUser
      summary: Unblock a user
      security:
        - bearerAuth: []
      responses:
        "204":
          description: User unblocked.
        "404":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /v1/reports:
    post:
      tags: [Safety]
      operationId: createReport
      summary: Report content or an account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReport"
      responses:
        "201":
          description: Report queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportEnvelope"
        "409":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "428":
          $ref: "#/components/responses/Error"
  /v1/copyright/notices:
    post:
      tags: [Copyright]
      operationId: createCopyrightNotice
      summary: Submit a copyright notice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCopyrightNotice"
      responses:
        "201":
          description: Notice received. Sensitive claimant data is not echoed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopyrightReceiptEnvelope"
        "422":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
...
  /v1/moderation/actions/{actionID}/appeals:
    parameters:
      - $ref: "#/components/parameters/ActionID"
    post:
      tags: [Safety]
      operationId: createAppeal
      summary: Appeal a moderation action
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 20
                  maxLength: 2000
      responses:
        "201":
          description: Appeal received.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppealEnvelope"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PolicyKind:
      name: kind
      in: path
      required: true
      schema:
        type: string
...
    UserID:
      name: userID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ActionID:
      name: actionID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
...
    Policy:
      type: object
      required: [kind, version, title, content, requires_acceptance, effective_at]
      properties:
        kind:
          type: string
        version:
          type: string
        title:
          type: string
        content:
          type: string
          description: Markdown policy content.
        requires_acceptance:
          type: boolean
        effective_at:
          type: string
          format: date-time
    PolicyEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Policy"
...
    CreateReport:
      type: object
      required: [target_type, target_id, reason]
      properties:
        target_type:
          type: string
          enum: [submission, user, challenge]
        target_id:
          type: string
          format: uuid
        reason:
          type: string
          enum:
            - harassment
            - hate
            - sexual_content
            - violence
            - self_harm
            - privacy
            - spam
            - cheating
            - copyright
            - underage
            - child_safety
            - other
        details:
          type: string
          maxLength: 2000
    Report:
      type: object
      required:
        - id
        - reporter_id
        - target_type
        - target_id
        - reason
        - status
        - priority
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        reporter_id:
          type: string
          format: uuid
        target_type:
          type: string
        target_id:
          type: string
          format: uuid
        reason:
          type: string
        details:
          type: string
        status:
          type: string
          enum: [open, reviewing, resolved, dismissed]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        assigned_to:
          type: [string, "null"]
          format: uuid
        resolution_note:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        resolved_at:
          type: [string, "null"]
          format: date-time
    ReportEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Report"
...
    Appeal:
      type: object
      required: [id, user_id, action_id, reason, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        action_id:
          type: string
          format: uuid
        reason:
          type: string
        status:
          type: string
          enum: [pending, upheld, reversed]
        reviewed_by:
          type: [string, "null"]
          format: uuid
        resolution_note:
          type: string
        created_at:
          type: string
          format: date-time
        resolved_at:
          type: [string, "null"]
          format: date-time
    AppealEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Appeal"
...
    CreateCopyrightNotice:
      type: object
      required:
        - claimant_name
        - claimant_email
        - claimant_address
        - relationship
        - copyrighted_work
        - infringing_url
        - good_faith
        - accuracy
        - signature
      properties:
        claimant_name:
          type: string
        claimant_email:
          type: string
          format: email
        claimant_address:
          type: string
        relationship:
          type: string
        copyrighted_work:
          type: string
        infringing_url:
          type: string
          format: uri
        submission_id:
          type: [string, "null"]
          format: uuid
        good_faith:
          type: boolean
          const: true
        accuracy:
          type: boolean
          const: true
        signature:
          type: string
...
    CopyrightReceiptEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [id, status, created_at]
          properties:
            id:
              type: string
              format: uuid
            notice_id:
              type: string
              format: uuid
            status:
              type: string
            created_at:
              type: string
              format: date-time
...